Cleaned up the Login Screen
Cleaned up the code of the loginSscreen.
Renaming the checkboxState state variable to rememberMe for clarity.
Simplifying the logic inside the useEffect hook.
Renaming the value variable to data inside the fetchData function.
Renaming the handleLogin function for clarity.
Renaming the isChecked prop of the BouncyCheckbox component to rememberMe.
Cleaning up unnecessary comments and debugging statements.