Timer bound to rest time
Now the timer will correspond to the resting time. If you rest too much it will keep counting with the string "extra" by the side of the timer. The extra time contain only the extra time itself. The timer now goes from rest time to 0. And from 0 to infinity, on the extra time.